Move relay, fix rebase issues
