+++ protocol/dialer.go
@@ -1,0 +1,11 @@
+package protocol
+
+import (
+	"context"
+
+	connection "github.com/nimona/go-nimona-fabric/connection"
+)
+
+type Dialer interface {
+	DialContext(ctx context.Context, as string) (context.Context, connection.Conn, error)
+}
